+++ compilation/src/main/12.2.dbr-scala/org/apache/spark/sql/catalyst/expressions/Expressions.scala
@@ -17,6 +17,9 @@ import org.apache.spark.sql.types._
 // taken from OSS master https://github.com/apache/spark/commit/cd32c22c9373333d2bd3b89a4ffae1b549396658
 abstract class Expression extends TreeNode[Expression] {
 
+  def stateful: Boolean = false
+  def freshCopyIfContainsStatefulExpression(): Expression = this
+
   /**
    * Returns true when an expression is a candidate for static evaluation before the query is
    * executed. A typical use case: [[org.apache.spark.sql.catalyst.optimizer.ConstantFolding]]