+++ runtime/src/main/3.4.1-scala/com/sparkutils/shim/expressions/Compat.scala
@@ -1,12 +1,16 @@
 package com.sparkutils.shim.expressions
 
 import org.apache.spark.sql.catalyst.expressions.{HigherOrderFunction, Nondeterministic, Unevaluable}
+// DBR 15.4 added nonVolatile
+trait NondeterministicLike extends Nondeterministic {
 
+}
 trait StatefulLike extends Nondeterministic {
   /**
    * Return a fresh uninitialized copy of the stateful expression.
    */
   def freshCopy(): StatefulLike
+  override def stateful: Boolean = true
 
 }
 
@@ -17,3 +21,6 @@ trait HigherOrderFunctionLike extends HigherOrderFunction {}
  */
 trait FoldableUnevaluable extends Unevaluable {
 }
+
+// dropped in 4.0
+trait NullIntolerant extends org.apache.spark.sql.catalyst.expressions.NullIntolerant