+++ docs/background/changelog.md
@@ -1,4 +1,14 @@
-### 0.0.1 <small>20th March, 2024</small>
+### 0.2.0 <small>11th June, 2025</small>
+
+#5 - Spark 4 support, including DBR 17.0
+
+#8 - DBR 16.4 support
+
+#9 - isStateful function and fixes
+
+#11 - copyStateful function
+
+### 0.0.1 <small>8th March, 2024</small>
 
 #1 - Quality support